feature(util/random): generate random bytes
We need this feature to send random bytes over the wire.

Compared to the github.com/neubot/botticelli implementation, this one
is more modular and most likely also more efficient.

Unlike botticelli, here we use directly an unlocked random source, which
is the fastest way of getting randomness.

There are two possibilities: generate random A-Za-z letters (what NDT
really needs) at maximum speed, and generate random bytes in a specific
set of bytes using a much slower generator algorithm.

The original copyright (which I mostly own) is gone since in the end
this is an implementation from scratch of a similar concept.

This commit also adds docstrings and (passing) tests.