refactor(protocol.go): sniff for WebSockets during Login
I would argue that we don't need to sniff for WebSockets at every
message, rather we want to do that only when reading Login.

In fact, as far as I understand, we should pick a protocol at the
beginning, and then stick to it for the rest of the session.

As such, it only matters to sniff for WebSockets during Login.

Added also a TODO comment explaining how I am planning to go about
further specializing the code next, so Greg can comment on it.